4045: Rename place files
