+++ native/src/assets/index.ts
@@ -11,7 +11,7 @@ import TuNewsInactiveIcon from '../../../assets/icons/tu-news-inactive.svg'
 import IntroLanguageIcon from '../../../assets/intro/language.svg'
 import IntroNewsIcon from '../../../assets/intro/news.svg'
 import IntroOfflineIcon from '../../../assets/intro/offline.svg'
-import IntroPoisIcon from '../../../assets/intro/pois.svg'
+import IntroPoisIcon from '../../../assets/intro/places.svg'
 import IntroSearchIcon from '../../../assets/intro/search.svg'
 import IntroWelcomeIcon from '../../../assets/intro/welcome.svg'
