+++ native/src/components/PlaceDetailRow.tsx
