+++ native/src/components/PlaceListItem.tsx
