+++ native/src/routes/Places.tsx
@@ -10,8 +10,8 @@ import { PoiCategoryModel, PoiModel, RegionModel } from 'shared/api'
 
 import { EditLocationIcon } from '../assets'
 import MapView from '../components/MapView'
-import PoiFiltersModal from '../components/PoiFiltersModal'
-import PoisBottomSheet from '../components/PoisBottomSheet'
+import PoiFiltersModal from '../components/PlaceFiltersModal'
+import PoisBottomSheet from '../components/PlacesBottomSheet'
 import Icon from '../components/base/Icon'
 import Text from '../components/base/Text'
 import dimensions from '../constants/dimensions'
