+++ shared/api/endpoints/createPlacesEndpoint.ts
@@ -8,8 +8,8 @@ import ContactModel from '../models/ContactModel'
 import LocationModel from '../models/LocationModel'
 import OpeningHoursModel from '../models/OpeningHoursModel'
 import OrganizationModel from '../models/OrganizationModel'
-import PoiCategoryModel from '../models/PoiCategoryModel'
-import PoiModel from '../models/PoiModel'
+import PoiCategoryModel from '../models/PlaceCategoryModel'
+import PoiModel from '../models/PlaceModel'
 import { JsonOpeningHoursType, JsonPoiType } from '../types'
 
 export const POIS_ENDPOINT_NAME = 'pois'
