+++ shared/api/endpoints/testing/index.ts
@@ -3,4 +3,4 @@ export { default as RegionModelBuilder } from './RegionModelBuilder'
 export { default as EventModelBuilder } from './EventModelBuilder'
 export { default as LanguageModelBuilder } from './LanguageModelBuilder'
 export { default as NewsModelBuilder } from './NewsModelBuilder'
-export { default as PoiModelBuilder } from './PoiModelBuilder'
+export { default as PoiModelBuilder } from './PlaceModelBuilder'
