+++ shared/utils/geoJson.ts
@@ -1,6 +1,6 @@
 import distance from '@turf/distance'
 
-import PoiModel from '../api/models/PoiModel'
+import PoiModel from '../api/models/PlaceModel'
 import { featureLayerId, MapFeature, MapFeatureCollection } from '../constants/map'
 
 export const MIN_DISTANCE_THRESHOLD = 0.1
