+++ web/src/components/PlaceChips.tsx
