+++ web/src/components/PlaceDetails.tsx
@@ -14,7 +14,7 @@ import { PoiModel } from 'shared/api'
 import Contact from './Contact'
 import CustomThumbnail from './CustomThumbnail'
 import OpeningHours from './OpeningHours'
-import PoiChips from './PoiChips'
+import PoiChips from './PlaceChips'
 import RemoteContent from './RemoteContent'
 import Accordion from './base/Accordion'
 import Link from './base/Link'
