+++ web/src/components/PlaceListItem.tsx
