+++ web/src/components/PlacePanelHeader.tsx
