+++ web/src/components/PlacePanelNavigation.tsx
