+++ web/src/components/__tests__/PlacesMobile.spec.tsx
@@ -4,7 +4,7 @@ import { LocationType, MapViewViewport, prepareMapFeatures } from 'shared'
 import { PoiModel, PoiModelBuilder } from 'shared/api'
 
 import { renderWithRouterAndTheme } from '../../testing/render'
-import PoisMobile from '../PoisMobile'
+import PoisMobile from '../PlacesMobile'
 
 jest.mock('react-i18next')
 jest.mock('../MapView', () => () => <div>MapView</div>)
