+++ web/src/hooks/useLoadSearchDocuments.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react'
 
 import { prepareSearchDocuments } from 'shared'
-import { createCategoriesEndpoint, createEventsEndpoint, createPOIsEndpoint, ExtendedDocumentModel } from 'shared/api'
+import { createCategoriesEndpoint, createEventsEndpoint, createPlacesEndpoint, ExtendedDocumentModel } from 'shared/api'
 
 import useQueryFromEndpoint from './useQueryFromEndpoint'
 
@@ -26,7 +26,7 @@ const useLoadSearchDocuments = ({
 
   const categories = useQueryFromEndpoint(createCategoriesEndpoint, cmsApiBaseUrl, params)
   const events = useQueryFromEndpoint(createEventsEndpoint, cmsApiBaseUrl, params)
-  const pois = useQueryFromEndpoint(createPOIsEndpoint, cmsApiBaseUrl, params)
+  const pois = useQueryFromEndpoint(createPlacesEndpoint, cmsApiBaseUrl, params)
 
   const documents = useMemo(
     () => prepareSearchDocuments(categories.data, events.data, pois.data),
